requirements.txt
# Core libraries (essential for basic functionality)
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
markdown>=3.4.0
pygments>=2.13.0
beautifulsoup4>=4.11.0  # Used for HTML parsing in course validator

# Course validator utilities
colorama>=0.4.6      # Terminal text formatting
tabulate>=0.9.0      # Pretty-printing tables
markdown>=3.4.0      # Markdown parsing

# Security
pyjwt>=2.7.0
passlib>=1.7.4
cryptography>=40.0.0
bcrypt>=4.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system=="Windows"  # Windows-specific binary

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5  # PostgreSQL driver (binary version, no build tools needed)
alembic>=1.10.0

# Data processing and analysis (pre-built wheels available for these)
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
pyyaml>=6.0.0
requests>=2.28.0

# =================================================================
# ADVANCED DEPENDENCIES (optional)
# The following packages may require build tools
# All core functionality will work without these packages
# =================================================================

# Deep learning (optional - binary wheels available for most platforms)
# If you have trouble installing these, you can install them separately with:
# pip install torch transformers -f https://download.pytorch.org/whl/torch_stable.html
torch>=2.0.0
transformers>=4.27.0

# Sentence embeddings (optional)
# If you have trouble, try: pip install sentence-transformers --no-deps
#sentence-transformers>=2.2.0

# Datasets (optional)
datasets>=2.11.0

# NLP libraries (optional)
# For spaCy, if you have trouble, install separately with: python -m spacy download en_core_web_sm
spacy>=3.5.0
tokenizers>=0.13.0

# Vector databases (optional)
faiss-cpu>=1.7.0
hnswlib>=0.7.0
annoy>=1.17.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Other useful utilities
rich>=13.0.0  # Rich text and formatting in terminal
pillow>=9.5.0  # Python Imaging Library

# =================================================================
# INSTALLATION GUIDE
# =================================================================
# If you encounter issues installing any package:
# 
# 1. For Windows users:
#    - Make sure you have Microsoft Visual C++ Build Tools installed
#      Download from: https://visualstudio.microsoft.com/visual-cpp-build-tools/
#    - For PyTorch: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# 2. For Mac users:
#    - Install Xcode command line tools: xcode-select --install
#
# 3. For Linux users:
#    - Install build essentials: apt-get install build-essential
#
# 4. Try installing just the core packages first:
#    - pip install fastapi uvicorn pydantic python-dotenv
#
# The run.py script will automatically handle installing essential packages. 